+++ GritKeeper/source/MainForm.cs
@@ -41,7 +41,7 @@ public partial class MainForm : Form
     ToolStripMenuItem undoMenuItem, redoMenuItem;
     ToolStripButton undoStatusBtn, redoStatusBtn;
 
-    internal const string AppVersion = "1.7.0";
+    internal const string AppVersion = "1.8.0";
 
     public MainForm()
     {